Test: 테스트에도 팀 초대 변경 사항을 반영함
