+++ src/test/java/com/growup/pms/docs/TeamInvitationControllerV1DocsTest.java
@@ -6,17 +6,16 @@ import static com.epages.restdocs.apispec.ResourceDocumentation.resource;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.epages.restdocs.apispec.SimpleType;
-import com.growup.pms.invitation.controller.dto.request.TeamInvitationCreateRequest;
-import com.growup.pms.invitation.service.TeamInvitationService;
-import com.growup.pms.invitation.service.dto.TeamInvitationCreateCommand;
+import com.growup.pms.role.domain.TeamRole;
+import com.growup.pms.team.controller.dto.request.TeamInvitationCreateRequest;
+import com.growup.pms.team.service.TeamInvitationService;
+import com.growup.pms.team.service.dto.TeamInvitationCreateCommand;
 import com.growup.pms.test.annotation.AutoKoreanDisplayName;
 import com.growup.pms.test.annotation.WithMockSecurityUser;
 import com.growup.pms.test.support.ControllerSliceTestSupport;
@@ -40,26 +39,26 @@ class TeamInvitationControllerV1DocsTest extends ControllerSliceTestSupport {
         // given
         Long 초대할_팀_ID = 1L;
         Long 초대할_사용자_ID = 1L;
-        Long 예상하는_생성된_초대_ID = 1L;
-        TeamInvitationCreateRequest 팀_초대_요청 = new TeamInvitationCreateRequest(초대할_사용자_ID);
+        String 초대된_역할명 = TeamRole.LEADER.getRoleName();
+        TeamInvitationCreateRequest 팀_초대_요청 = new TeamInvitationCreateRequest(초대할_사용자_ID, 초대된_역할명);
 
-        when(teamInvitationService.sendInvitation(eq(초대할_팀_ID), any(TeamInvitationCreateCommand.class))).thenReturn(예상하는_생성된_초대_ID);
+        doNothing().when(teamInvitationService).sendInvitation(eq(초대할_팀_ID), any(TeamInvitationCreateCommand.class));
 
         // when & then
         mockMvc.perform(post("/api/v1/team/{teamId}/invitation", 초대할_팀_ID)
                 .content(objectMapper.writeValueAsString(팀_초대_요청))
                 .contentType(MediaType.APPLICATION_JSON)
                 .header(HttpHeaders.AUTHORIZATION, "Bearer 액세스 토큰"))
-                .andExpectAll(
-                        status().isCreated(),
-                        header().string(HttpHeaders.LOCATION, "/api/v1/team/1/invitation/1"))
+                .andExpect(status().isOk())
                 .andDo(docs.document(resource(
                         ResourceSnippetParameters.builder()
                                 .tag(TAG)
                                 .summary("팀 초대")
                                 .description("해당 사용자를 특정 팀에 초대합니다.")
                                 .pathParameters(parameterWithName("teamId").type(SimpleType.NUMBER).description("초대할 팀 ID"))
-                                .requestFields(fieldWithPath("userId").type(JsonFieldType.NUMBER).description("초대할 사용자 ID"))
+                                .requestFields(
+                                        fieldWithPath("userId").type(JsonFieldType.NUMBER).description("초대할 사용자 ID"),
+                                        fieldWithPath("roleName").type(JsonFieldType.STRING).description("초대할 사용자의 역할명"))
                                 .requestHeaders(
                                         headerWithName(HttpHeaders.CONTENT_TYPE).type(SimpleType.STRING).description(MediaType.APPLICATION_JSON_VALUE),
                                         headerWithName(HttpHeaders.AUTHORIZATION).type(SimpleType.STRING).description("Bearer 액세스 토큰"))
@@ -77,7 +76,7 @@ class TeamInvitationControllerV1DocsTest extends ControllerSliceTestSupport {
         // when & then
         mockMvc.perform(post("/api/v1/team/{teamId}/invitation/accept", 초대된_팀_ID)
                         .header(HttpHeaders.AUTHORIZATION, "Bearer 액세스 토큰"))
-                .andExpect(status().isNoContent())
+                .andExpect(status().isOk())
                 .andDo(docs.document(resource(
                         ResourceSnippetParameters.builder()
                                 .tag(TAG)
@@ -99,7 +98,7 @@ class TeamInvitationControllerV1DocsTest extends ControllerSliceTestSupport {
         // when & then
         mockMvc.perform(post("/api/v1/team/{teamId}/invitation/decline", 초대된_팀_ID)
                         .header(HttpHeaders.AUTHORIZATION, "Bearer 액세스 토큰"))
-                .andExpect(status().isNoContent())
+                .andExpect(status().isOk())
                 .andDo(docs.document(resource(
                         ResourceSnippetParameters.builder()
                                 .tag(TAG)
